creating 'contains' feature
