--- src/main.ts
@@ -1,14 +1,0 @@
-import LinkedList from "./LinkedList";
-
-const dummyList = new LinkedList<number>();
-
-dummyList.append(1);
-dummyList.append(2);
-dummyList.append(3);
-dummyList.append(4);
-
-const iterator = dummyList.traverse();
-
-for(const value of iterator) {
-    console.log(value);
-}
